Refactor footer
