+++ src/pages/contactUs/contactUsComp/ContactUsComp.jsx
@@ -1,4 +1,4 @@
-import Footer from "../../../components/Footer";
+import {Footer} from "../../../components";
 import Navbar from "../../../components/Navbar";
 import ContactUsDetails from "../contactDetails/ContactUsDetails";
 import ContactUsForm from "../contactUsForm/ContactUsForm";
