+++ src/pages/external/policy/index.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useRef } from "react";
-import Footer from "../../../components/Footer";
+import {Footer} from "../../../components";
 import Navbar from "../../../components/Navbar";
 import { contentsData,  policyData } from "./policyDatas";
 import DOMPurity from 'dompurify'
